+++ StatisticsTests/Histograms/ThreadsafeHistogramTests.cs
@@ -286,7 +286,7 @@ namespace StatisticsTests.Histograms
             double z = stdNormal.InverseCDF(.5 + .5 * .85);
             var convergencecriteria = new ConvergenceCriteria(maxIterations: maxiter, tolerance: .1, zAlpha: z);
             ThreadsafeInlineHistogram histogram = new ThreadsafeInlineHistogram(convergencecriteria);
-            int iterations = convergencecriteria.MinIterations;
+            Int64 iterations = convergencecriteria.MinIterations;
             object whilelock = new object();
             while (!histogram.IsConverged)
             {
@@ -333,8 +333,8 @@ namespace StatisticsTests.Histograms
             histogram.ForceDeQueue();
             XElement xElement = histogram.WriteToXML();
             ThreadsafeInlineHistogram histogramFromXML = ThreadsafeInlineHistogram.ReadFromXML(xElement);
-            int[] expectedBinCounts = histogram.BinCounts;
-            int[] actualBinCounts = histogramFromXML.BinCounts;
+            Int64[] expectedBinCounts = histogram.BinCounts;
+            Int64[] actualBinCounts = histogramFromXML.BinCounts;
             Assert.Equal(expectedBinCounts, actualBinCounts);
         }
     }
